requirements.txt
PyQt5=5.9.2
request=2.13.0